fix: moved len_tokens and msgs2dicts from util.py to message.py
